#155 separate workspace parsing logic
